requirements.txt
langchain
langchain-community
faiss-cpu
transformers
torch
unstructured
docx2txt
pypdf
sentence-transformers
fpdf
streamlit
fastapi
uvicorn
python-multipart
passlib[bcrypt]
sqlalchemy
python-jose[cryptography]
python-dotenv
pydantic[email]
